Eliminó CaificarPlato.java q no hacía nada, al layout detalle plato le a...

+++ app/src/main/java/com/rejuntadosdeinge/umenu/modelo/Plato.java
@@ -15,7 +15,7 @@ public class Plato {
     private String tipo;
     private int calificaciones;
     private int total;
-    private int promedio;
+    private float promedio;
 
     public int getId() {
         return id;
@@ -66,9 +66,7 @@ public class Plato {
     public void setTotal(int total) {
         this.total = total;
     }
-    public int getPromedio() {
-        return promedio;
-    }
+    public float getPromedio() { return promedio; }
     public void setPromedio(int promedio) {
         this.promedio = promedio;
     }
